🐜 fix: Fix an issue where ts-server is loading forever
