+++ packages/functions/README.md
@@ -1,4 +1,4 @@
-# auth
+# @nestri/functions
 
 To install dependencies:
 
@@ -12,4 +12,4 @@ To run:
 bun run index.ts
 ```
 
-This project was created using `bun init` in bun v1.1.34. [Bun](https://bun.sh) is a fast all-in-one JavaScript runtime.
+This project was created using `bun init` in bun v1.2.11. [Bun](https://bun.sh) is a fast all-in-one JavaScript runtime.
