+++ packages/functions/src/api/utils/hook.ts
