+++ packages/functions/src/auth/adapters/github.ts
@@ -1,4 +1,4 @@
-import { Oauth2Adapter, type Oauth2WrappedConfig } from "./oauth2"
+import { Oauth2Adapter, type Oauth2WrappedConfig } from "../ui/oauth2"
 
 export function GithubAdapter(config: Oauth2WrappedConfig) {
   return Oauth2Adapter({
