+++ packages/functions/src/auth/adapters/index.ts
@@ -1,0 +1,3 @@
+export * from "./discord"
+export * from "./github"
+export * from "./password"
