+++ packages/functions/src/auth/index.ts
@@ -1,19 +1,15 @@
 import { Resource } from "sst"
-import { Select } from "./ui/select";
-import { subjects } from "./subjects"
 import { logger } from "hono/logger";
-import { PasswordUI } from "./ui/password"
-import { patchLogger } from "./log-polyfill";
+import { subjects } from "../subjects"
+import { Select, PasswordUI } from "./ui";
 import { issuer } from "@openauthjs/openauth";
 import { User } from "@nestri/core/user/index"
-import { Email } from "@nestri/core/email/index";
-import { handleDiscord, handleGithub } from "./utils";
-import { GithubAdapter } from "./ui/adapters/github";
-import { Machine } from "@nestri/core/machine/index"
-import { DiscordAdapter } from "./ui/adapters/discord";
-import { PasswordAdapter } from "./ui/adapters/password";
-import { type Provider } from "@openauthjs/openauth/provider/provider"
+// import { Email } from "@nestri/core/email/index";
+// import { Machine } from "@nestri/core/machine/index"
+import { handleDiscord, handleGithub, patchLogger } from "./utils";
 import { MemoryStorage } from "@openauthjs/openauth/storage/memory";
+// import { type Provider } from "@openauthjs/openauth/provider/provider"
+import { DiscordAdapter, PasswordAdapter, GithubAdapter } from "./adapters";
 
 type OauthUser = {
     primary: {
@@ -80,25 +76,25 @@ const app = issuer({
                 },
             }),
         ),
-        machine: {
-            type: "machine",
-            async client(input) {
-                // FIXME: Do we really need this?
-                // if (input.clientSecret !== Resource.AuthFingerprintKey.value) {
-                //     throw new Error("Invalid authorization token");
-                // }
+        // machine: {
+        //     type: "machine",
+        //     async client(input) {
+        //         // FIXME: Do we really need this?
+        //         // if (input.clientSecret !== Resource.AuthFingerprintKey.value) {
+        //         //     throw new Error("Invalid authorization token");
+        //         // }
 
-                const fingerprint = input.params.fingerprint;
-                if (!fingerprint) {
-                    throw new Error("Hostname is required");
-                }
+        //         const fingerprint = input.params.fingerprint;
+        //         if (!fingerprint) {
+        //             throw new Error("Hostname is required");
+        //         }
 
-                return {
-                    fingerprint,
-                };
-            },
-            init() { }
-        } as Provider<{ fingerprint: string; }>,
+        //         return {
+        //             fingerprint,
+        //         };
+        //     },
+        //     init() { }
+        // } as Provider<{ fingerprint: string; }>,
     },
     allow: async (input) => {
         const url = new URL(input.redirectURI);
@@ -108,39 +104,40 @@ const app = issuer({
         return false;
     },
     success: async (ctx, value, req) => {
-        if (value.provider === "machine") {
-            const countryCode = req.headers.get('CloudFront-Viewer-Country') || 'Unknown'
-            const country = req.headers.get('CloudFront-Viewer-Country-Name') || 'Unknown'
-            const latitude = Number(req.headers.get('CloudFront-Viewer-Latitude')) || 0
-            const longitude = Number(req.headers.get('CloudFront-Viewer-Longitude')) || 0
-            const timezone = req.headers.get('CloudFront-Viewer-Time-Zone') || 'Unknown'
-            const fingerprint = value.fingerprint
+        // I dunno what i broke... will check later
+        // if (value.provider === "machine") {
+        //     const countryCode = req.headers.get('CloudFront-Viewer-Country') || 'Unknown'
+        //     const country = req.headers.get('CloudFront-Viewer-Country-Name') || 'Unknown'
+        //     const latitude = Number(req.headers.get('CloudFront-Viewer-Latitude')) || 0
+        //     const longitude = Number(req.headers.get('CloudFront-Viewer-Longitude')) || 0
+        //     const timezone = req.headers.get('CloudFront-Viewer-Time-Zone') || 'Unknown'
+        //     const fingerprint = value.fingerprint
 
-            const existing = await Machine.fromFingerprint(fingerprint)
-            if (!existing) {
-                const machineID = await Machine.create({
-                    countryCode,
-                    country,
-                    fingerprint,
-                    timezone,
-                    location: {
-                        latitude,
-                        longitude
-                    },
-                    //FIXME: Make this better
-                    // userID: null
-                })
-                return ctx.subject("machine", {
-                    machineID,
-                    fingerprint
-                });
-            }
+        //     const existing = await Machine.fromFingerprint(fingerprint)
+        //     if (!existing) {
+        //         const machineID = await Machine.create({
+        //             countryCode,
+        //             country,
+        //             fingerprint,
+        //             timezone,
+        //             location: {
+        //                 latitude,
+        //                 longitude
+        //             },
+        //             //FIXME: Make this better
+        //             // userID: null
+        //         })
+        //         return ctx.subject("machine", {
+        //             machineID,
+        //             fingerprint
+        //         });
+        //     }
 
-            return ctx.subject("machine", {
-                machineID: existing.id,
-                fingerprint
-            });
-        }
+        //     return ctx.subject("machine", {
+        //         machineID: existing.id,
+        //         fingerprint
+        //     });
+        // }
 
         // TODO: This works, so use this while registering the task
         // console.log("country_code", req.headers.get('CloudFront-Viewer-Country'))
