+++ packages/functions/src/auth/ui/oauth2.tsx
@@ -1,5 +1,6 @@
 /** @jsxImportSource hono/jsx */
-import { Layout } from "../base"
+import fetch from "node-fetch"
+import { Layout } from "./base"
 import { OauthError } from "@openauthjs/openauth/error"
 import { getRelativeUrl } from "@openauthjs/openauth/util"
 import { type Provider } from "@openauthjs/openauth/provider/provider"
