--- packages/functions/src/type.ts
@@ -1,4 +1,0 @@
-export enum Subscription {
-  Pro = "Pro",
-  Free = "Free"
-}
